# api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cursos": {
      "post": {
        "operationId": "CursosController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarCursoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "O curso foi criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursoRespostaDto"
                }
              }
            }
          },
          "409": {
            "description": "Já existe um curso cadastrado com este nome."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao criar o curso."
          }
        },
        "summary": "Criar um novo curso",
        "tags": [
          "Cursos"
        ]
      },
      "get": {
        "operationId": "CursosController_buscarTodos",
        "parameters": [
          {
            "name": "pagina",
            "required": false,
            "in": "query",
            "description": "Número da página atual (padrão: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "itensPorPagina",
            "required": false,
            "in": "query",
            "description": "Número de itens por página (padrão: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta paginada recebida com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginacaoRespostaDto"
                    },
                    {
                      "properties": {
                        "dados": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CursoRespostaDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Os parâmetros de paginação (página ou itensPorPagina) devem ser maiores ou iguais a 1."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar os cursos."
          }
        },
        "summary": "Buscar uma lista paginada de cursos",
        "tags": [
          "Cursos"
        ]
      }
    },
    "/cursos/{id}": {
      "get": {
        "operationId": "CursosController_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "O curso foi encontrado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursoRespostaDto"
                }
              }
            }
          },
          "404": {
            "description": "Curso com o ID especificado não encontrado."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar o curso."
          }
        },
        "summary": "Buscar curso pelo ID",
        "tags": [
          "Cursos"
        ]
      },
      "patch": {
        "operationId": "CursosController_atualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarCursoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "O curso foi atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursoRespostaDto"
                }
              }
            }
          },
          "404": {
            "description": "Curso não encontrado."
          },
          "409": {
            "description": "Já existe um curso cadastrado com este nome."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao atualizar o curso."
          }
        },
        "summary": "Atualizar curso pelo ID",
        "tags": [
          "Cursos"
        ]
      },
      "delete": {
        "operationId": "CursosController_remover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "O curso foi deletado com sucesso."
          },
          "404": {
            "description": "Curso não encontrado."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao deletar o curso."
          }
        },
        "summary": "Deletar curso pelo ID",
        "tags": [
          "Cursos"
        ]
      }
    },
    "/planos-curso": {
      "post": {
        "operationId": "PlanosCursoController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarPlanoCursoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "O plano de curso foi criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoCursoRespostaDto"
                }
              }
            }
          },
          "409": {
            "description": "Já existe um plano de curso cadastrado com este nome."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao criar o plano de curso."
          }
        },
        "summary": "Criar um novo plano de curso",
        "tags": [
          "Planos de Curso"
        ]
      },
      "get": {
        "operationId": "PlanosCursoController_buscarTodos",
        "parameters": [
          {
            "name": "pagina",
            "required": false,
            "in": "query",
            "description": "Número da página atual (padrão: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "itensPorPagina",
            "required": false,
            "in": "query",
            "description": "Número de itens por página (padrão: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "cursoId",
            "required": false,
            "in": "query",
            "description": "Filtra os planos de curso pertencentes a um curso específico",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta paginada recebida com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginacaoRespostaDto"
                    },
                    {
                      "properties": {
                        "dados": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PlanoCursoRespostaDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Os parâmetros de paginação (página ou itensPorPagina) devem ser maiores ou iguais a 1."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar os planos de curso."
          }
        },
        "summary": "Buscar uma lista paginada de planos de curso",
        "tags": [
          "Planos de Curso"
        ]
      }
    },
    "/planos-curso/{id}": {
      "get": {
        "operationId": "PlanosCursoController_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "O plano de curso foi encontrado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoCursoRespostaDto"
                }
              }
            }
          },
          "404": {
            "description": "Plano de curso com o ID especificado não encontrado."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao buscar o plano de curso."
          }
        },
        "summary": "Buscar um plano de curso pelo ID",
        "tags": [
          "Planos de Curso"
        ]
      },
      "patch": {
        "operationId": "PlanosCursoController_atualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarPlanoCursoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "O plano de curso foi atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoCursoRespostaDto"
                }
              }
            }
          },
          "404": {
            "description": "Plano de curso com o ID especificado não encontrado."
          },
          "409": {
            "description": "Já existe um plano de curso cadastrado com este nome."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao atualizar o plano de curso."
          }
        },
        "summary": "Atualizar um plano de curso pelo ID",
        "tags": [
          "Planos de Curso"
        ]
      },
      "delete": {
        "operationId": "PlanosCursoController_remover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "O plano de curso foi removido com sucesso."
          },
          "404": {
            "description": "Plano de curso com o ID especificado não encontrado."
          },
          "500": {
            "description": "Ocorreu um erro inesperado ao remover o plano de curso."
          }
        },
        "summary": "Deletar um plano de curso pelo ID",
        "tags": [
          "Planos de Curso"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portal ONG Ser Amor API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginacaoMetaDto": {
        "type": "object",
        "properties": {
          "paginaAtual": {
            "type": "number",
            "description": "Página atual",
            "example": 1
          },
          "itensPorPagina": {
            "type": "number",
            "description": "Número de itens por página",
            "example": 10
          },
          "totalItens": {
            "type": "number",
            "description": "Total de itens",
            "example": 100
          },
          "totalPaginas": {
            "type": "number",
            "description": "Total de páginas",
            "example": 10
          }
        },
        "required": [
          "paginaAtual",
          "itensPorPagina",
          "totalItens",
          "totalPaginas"
        ]
      },
      "PaginacaoRespostaDto": {
        "type": "object",
        "properties": {
          "meta": {
            "description": "Metadados da paginação",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginacaoMetaDto"
              }
            ]
          },
          "dados": {
            "description": "Dados paginados",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "meta",
          "dados"
        ]
      },
...
      "CriarCursoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Avenida Brasil"
          }
        },
        "required": [
          "nome"
        ]
      },
      "CursoRespostaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "nome": {
            "type": "string",
            "example": "Informática Básica"
          }
        },
        "required": [
          "id",
          "nome"
        ]
      },
      "AtualizarCursoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Avenida Brasil"
          }
        }
      },
      "CriarPlanoCursoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Introdução à Lógica de Programação"
          },
          "cursoId": {
            "type": "string",
            "example": "1"
          }
        },
        "required": [
          "nome",
          "cursoId"
        ]
      },
      "PlanoCursoRespostaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "nome": {
            "type": "string",
            "example": "Plano Básico de Informática"
          },
          "cursoId": {
            "type": "string",
            "example": "1"
          },
          "curso": {
            "$ref": "#/components/schemas/CursoRespostaDto"
          }
        },
        "required": [
          "id",
          "nome",
          "cursoId"
        ]
      },
      "AtualizarPlanoCursoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Introdução à Lógica de Programação"
          }
        }
      },
...
    }
  }
}